requirements.txt
aiohttp
python-dotenv
playwright
